refactor: extract ReviewWorkspaceTab.swift from ReviewWorkspaceTypes.swift
Splits the ReviewWorkspaceTab enum (UI navigation) out of
ReviewWorkspaceTypes.swift so timeline-entry / summary-group data
types remain grouped together.

Byte-identical move of lines 3-23. Only imports Foundation.

ReviewWorkspaceTypes.swift: 73 → 51 lines. New file 23 lines.
Tests: 667.
